+++ tests/Core/Spec/Spec14Test.php
@@ -24,6 +24,7 @@ declare(strict_types=1);
 namespace CycloneDX\Tests\Core\Spec;
 
 use CycloneDX\Core\Spec\Format;
+use CycloneDX\Core\Spec\Spec14;
 use CycloneDX\Core\Spec\SpecInterface;
 
 /**